crtxtenc: fix warning

+++ crengine/src/crtxtenc.cpp
@@ -2070,10 +2070,10 @@ void MakeStatsForFile( const char * fname, const char * cp_name, const char * la
    if (!in)
       return;
    fseek( in, 0, SEEK_END );
-   int buf_size = ftell(in);
+   long buf_size = ftell(in);
    fseek( in, 0, SEEK_SET );
    unsigned char * buf = new unsigned char[buf_size];
-   if ( fread(buf, 1, buf_size, in) != buf_size ) {
+   if ( (long)fread(buf, 1, buf_size, in) != buf_size ) {
        fclose(in);
        return;
    }
